requirements.txt
# Core Web Framework
streamlit>=1.28.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# PDF Processing & Text Extraction
PyMuPDF>=1.23.0
pdf2image>=3.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

# AI & Natural Language Processing
openai>=1.0.0
sentence-transformers>=2.2.0
spacy>=3.6.0
tiktoken>=0.5.0

# Machine Learning & Scientific Computing
scikit-learn>=1.3.0
scipy>=1.11.0

# Data Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Report Generation
xhtml2pdf>=0.2.11

# UI Components & Styling
streamlit-option-menu>=0.3.6

# Environment & Configuration Management
python-dotenv>=1.0.0

# Utilities
typing-extensions>=4.0.0

# spaCy language model
https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.6.0/en_core_web_md-3.6.0-py3-none-any.whl